+++ middleware/advancedResults.js
@@ -54,7 +54,7 @@ const advancedResults = (model, join, on) =>
         format
       );
     }
-    console.log(query.red);
+    //console.log(query.red);
 
     // // Copy req.query, if any
     const reqQuery = { ...req.query };
@@ -71,6 +71,7 @@ const advancedResults = (model, join, on) =>
     if (Object.keys(reqQuery).length !== 0) {
       filterQuery += 'WHERE ';
       for (let [key, value] of Object.entries(reqQuery)) {
+        
         const mappedFilter = {
           key,
           value: value.split(',').map(str => str.split(/'/).join('')), // remove all ' for pgp formatter to parse into sql