+++ routes/listing-routes.js
@@ -18,12 +18,14 @@ const { uploadFile } = require('../utils/fileUploader');
 const {
   getListings,
   getAllListingsOwnedByUser,
+  getListingsAll,
   getListing,
   getListingByHashId,
   createListing,
   updateListing,
   verifyListing,
   deleteListing,
+  deactivateListing,
   uploadListingPics,
 } = require('../controllers/listings');
 
@@ -35,6 +37,7 @@ const listingUpdateRoute = require('./listingupdate-routes');
 const milestoneRoute = require('./milestone-routes');
 const participantRoute = require('./participant-routes');
 const listingSkillRoute = require('./listingskill-routes');
+const jobRoute = require('./job-routes');
 const storyRoute = require('./listingstory-routes');
 
 // Re-route this URI to other resource router
@@ -46,11 +49,12 @@ router.use('/:listing_id/listing-updates', listingUpdateRoute);
 router.use('/:listing_id/milestones', milestoneRoute);
 router.use('/:listing_id/participants', participantRoute);
 router.use('/:listing_id/listing-skills', listingSkillRoute);
+router.use('/:listing_id/jobs', jobRoute);
 
 // map routes to controller
 router
   .route('/')
-  .get(advancedResults('listings'), getListings)
+  .get(advancedResults('listingsview'), getListings)
   .post(
     protect,
     uploadFile.array('pics', 5),
@@ -77,6 +81,7 @@ router
   );
 
 router.route('/owner').get(getAllListingsOwnedByUser);
+router.route('/all').get(protect, authorise('admin'), advancedResults('listings'), getListingsAll);
 router.route('/:id/raw').get(getListing);
 router.route('/:hashId').get(getListingByHashId);
 
@@ -124,6 +129,14 @@ router
     updateListing
   )
   .delete(protect, deleteListing);
+
+router
+  .route('/:id/deactivate')
+  .put(
+    protect,
+    authorise('admin', 'owner'),
+    deactivateListing
+  );
 
 router
   .route('/:id/photo')